payment_service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
boto3==1.34.162
python-dotenv==1.0.1
python-jose==3.3.0
pydantic==2.8.2
httpx==0.27.0